chore: add playwright test runner
This adds the playwright test runner to the project. This will allow
us to run tests in a headless browser, which will be useful for
testing the UI. The test runner is configured to run in chromium and
firefox.

Basic test is added to ensure that the test runner is working.